FEATURE(redis): add local backup of the database
 ##### SUMMARY

Allow to do some local backups for the redis database. As it is
replicated to other redis in the cluster, a local backup is enough.

Using the variable `openio_redis_backups` it is possible to define
multiple backups. See sample in `default/main.yml`.

By default, no backups are defined.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION