CHANGE(redis): Activate `no-appendfsync-on-rewrite`
 ##### SUMMARY

When the AOF fsync policy is set to always or everysec, and a background
saving process (a background save or AOF log background rewriting) is
performing a lot of I/O against the disk, in some Linux configurations
Redis may block too long on the fsync() call. Note that there is no fix for
this currently, as even performing fsync in a different thread will block
our synchronous write(2) call.

In order to mitigate this problem it's possible to use the following option
that will prevent fsync() from being called in the main process while a
BGSAVE or BGREWRITEAOF is in progress.

This means that while another child is saving, the durability of Redis is
the same as "appendfsync none". In practical terms, this means that it is
possible to lose up to 30 seconds of log in the worst scenario (with the
default Linux settings).

If you have latency problems turn this to "yes". Otherwise leave it as
"no" that is the safest pick from the point of view of durability.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION
OB-556